attempting to get a stacked array
stacked array contains flattened and normalised data to be used in the NN. Having issues with the shape of the stacked array, this is due to the images having different lengths when flattened causing the dimension of the stacked array to be 1. If the lengths of the flattened images were the same, then we would see a second dimension of the stacked array.